requirements.txt
alabaster==0.7.13
annotated-types==0.6.0
antlr4-python3-runtime==4.13.1
attrs==23.1.0
Babel==2.13.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.4
Deprecated==1.2.14
distlib==0.3.8
dnspython==2.5.0
docutils==0.20.1
email-validator==2.1.0.post1
fhir.resources==6.5.0
fhirclient==4.1.0
fhirpath==0.10.5
fhirpathpy==1.1.0
fhirspec==0.4.0
filelock==3.13.1
idna==3.6
imagesize==1.4.1
importlib-metadata==6.8.0
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.5
nose==1.3.7
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.1
pathling==6.4.0
platformdirs==4.2.0
pluggy==0.13.1
py==1.11.0
py4j==0.10.9.7
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.16.1
pyspark==3.4.2
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
shellingham==1.4.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
toml==0.10.2
typer==0.3.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
virtualenv==20.25.0
wrapt==1.15.0
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
